picoxcell/pc3x3: pm support
pc3x3 does not allow the CPU or memory interface clks to be disabled,
only the peripheral clks.  Enable low power mode where the peripheral
clks are gated and we wait for an external interrupt.  Allow the
watchdog to be paused when entering suspend mode so that long suspend
periods don't trigger a watchdog reboot.

Signed-off-by: Jamie Iles <jamie@jamieiles.com>