drivers/otp: add support for Picoxcell PC3X3 OTP
The OTP in PC3X3 is 16KB and can be split into 1, 2, 4 or 8
regions each taking a different redundancy format.  The formats
increase the redundancy of the memory at the expense of bit storage
to help protect against programming errors for devices in the field.

Changes since v3:
	- Move the read_word/write_word ops into device ops.

Changes since v2:
	- Add support for CONFIG_OTP_WRITE_ENABLE.

Signed-off-by: Jamie Iles <jamie@jamieiles.com>