macb: support for RX packet timestamping
GEM devices support hardware timestamping of ingress packets.  The
nanoseconds portion of the timestamp is stored in the FCS of the
packet after validation and the seconds portion is accessible from
registers in the GEM itself.

If the GEM is configured to support timestamping then we either use
the TSU input clock for providing the clock rate if supported or pclk
if not.

Signed-off-by: Jamie Iles <jamie@jamieiles.com>