picogpio: add an ioctl() based userspace gpio interface
The gpiolib infrastructure provides a userspace interface to the
gpio pins on a board but this is insufficient for some applications:

	- Pin ownership isn't bound to an application.
	- Some people don't like sysfs etc.

picogpio is a very thin wrapper around the gpiolib interface that
provides ioctl calls to access gpio pins. Ownership is bound to an
application so all pins will be released when the application
terminates.