mm/msync: tweak tmpfs patch for syscall msync
Commit 1c3ae5441 "mm: msync: fix issues of sys_msync on tmpfs"
fixes the problem that sys_msync fails with tmpfs on MIPS CPU which
has feature "cache alias".

But it makes POSIX test cases mlockall/3-6 3-7 fail on MIPS.
Case mlockall/3-6 creates a share memory, and maps it to memory.
  fd = shm_open(SHM_NAME, O_RDWR|O_CREAT, S_IRUSR|S_IWUSR);
  foo = mmap(NULL, BUF_SIZE, PROT_WRITE, MAP_SHARED, fd, 0);
Then calls mlockall to lock all of the virtual address space:
  if (mlockall(MCL_CURRENT) == -1) {
At last tests whether the virtual address spaces are locked:
  page_ptr = (void*) ((long)foo - ((long)foo % page_size));
  result = msync(page_ptr, page_size, MS_SYNC|MS_INVALIDATE);
It espects msync returns -1 with EBUSY but returns 0.
Case mlockall/3-7 creates a normal file to mmap, and it fails too.

Tweak the patch to:
1 Moved the CONFIG_TMPFS block down in the loop so that the normal
  vma flag checking will perform.
2 There may be other VMAs in the list after this VMA which belongs
  to tmpfs file, and we should sync them after the tmpfs one. So
  remove the break loop clauses.

Signed-off-by: Kang Kai <kai.kang@windriver.com>