mtd: enabled buffered programming for M29EW in x8 Mode
This patch fixes a problem related to an incorrect value contained
in the CFI of M29EW devices family. The incorrect CFI field is
MaxBufWriteSize that should be 0x8 if the device is used in 8bit mode, whereas the
value read out from CFI is 0xA.

This patch is documented in Micron App note TN-13-07.

This commit forms part of the resolution of Mantis Issue #10786